Move HydraCLI to different file
